refactor: abstract getBookmarks query
